Multi-selection in properties dialog: bugfixing and finalization

+++ src/layui/layui/layPropertiesDialog.h
@@ -86,7 +86,7 @@ private:
   db::Manager *mp_manager;
   lay::Editables *mp_editables;
   int m_index, m_prev_index;
-  std::vector<int> m_object_indexes;
+  std::vector<size_t> m_object_indexes;
   QStackedLayout *mp_stack;
   QLabel *mp_none;
   lay::MainWindow *mp_mw;
